renames functions for clarity
also removes unnecessary double bangs
