+++ test/ImmutableArray.spec.js
@@ -1,7 +1,7 @@
 var testCompat   = require("./ImmutableArray/test-compat.js");
 var testFlatMap  = require("./ImmutableArray/test-flatMap.js");
 var testAsObject = require("./ImmutableArray/test-asObject.js");
-var testToMutable = require("./ImmutableArray/test-toMutable.js");
+var testToMutable = require("./ImmutableArray/test-asMutable.js");
 
 describe("ImmutableArray", function() {
   testCompat();
