+++ test/ImmutableArray/test-asMutable.js
@@ -6,11 +6,11 @@ var _         = require("lodash")
 var check     = TestUtils.check;
 
 module.exports = function() {
-  describe("#toMutable", function() {
+  describe("#asMutable", function() {
     it("returns a shallow mutable copy if not provided the deep flag", function() {
       check(100, [ JSC.array([TestUtils.TraversableObjectSpecifier, JSC.any()]) ], function(array) {
         var immutable = Immutable(array);
-        var mutable = immutable.toMutable();
+        var mutable = immutable.asMutable();
         assert.isFalse( Immutable.isImmutable(mutable));
         assert.isTrue( Immutable.isImmutable(mutable[0]));
         assert.isTrue( Immutable.isImmutable(mutable[0]['deep']));
@@ -22,7 +22,7 @@ module.exports = function() {
     it("returns a deep mutable copy if provided the deep flag", function() {
       check(100, [ JSC.array([TestUtils.TraversableObjectSpecifier, JSC.any()]) ], function(array) {
         var immutable = Immutable(array);
-        var mutable = immutable.toMutable({ deep: true });
+        var mutable = immutable.asMutable({ deep: true });
         assert.isFalse( Immutable.isImmutable(mutable));
         assert.isFalse( Immutable.isImmutable(mutable[0]));
         assert.isFalse( Immutable.isImmutable(mutable[0]['deep']));
