+++ test/ImmutableObject.spec.js
@@ -1,7 +1,7 @@
 var testMerge   = require("./ImmutableObject/test-merge.js");
 var testCompat  = require("./ImmutableObject/test-compat.js");
 var testWithout = require("./ImmutableObject/test-without.js");
-var testToMutable = require("./ImmutableArray/test-toMutable.js");
+var testToMutable = require("./ImmutableArray/test-asMutable.js");
 
 describe("ImmutableObject", function() {
   testCompat();
